third_party/blink/web_tests/platform/generic/fast/forms/input-minmax-expected.txt

Tests the behavior of .min and .max of HTMLInputElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.min is ""
PASS input.max is ""
PASS input.min is "0"
PASS input.max is "100"
PASS input.min is "abc"
PASS input.max is "qwerty"
PASS input.getAttribute("min") is "foo"
PASS input.getAttribute("max") is "bar"
PASS input.getAttribute("min") is ""
PASS input.getAttribute("max") is ""
Setting null to min:
PASS input.min is "null"
PASS input.getAttribute("min") is "null"
PASS input.min is "null"
Setting null to max:
PASS input.max is "null"
PASS input.getAttribute("max") is "null"
PASS input.max is "null"
Setting undefined to min:
PASS input.min is "undefined"
PASS input.getAttribute("min") is "undefined"
PASS input.min is "undefined"
Setting undefined to max:
PASS input.max is "undefined"
PASS input.getAttribute("max") is "undefined"
PASS input.max is "undefined"
Setting non-string to min:
PASS input.min is "256"
PASS input.getAttribute("min") is "256"
PASS input.min is "256"
Setting non-string to max:
PASS input.max is "256"
PASS input.getAttribute("max") is "256"
PASS input.max is "256"
Check implicit min/max of type=range:
PASS input.min is ""
PASS input.max is ""
PASS successfullyParsed is true

TEST COMPLETE
